# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple rate limiting and synchronization service for distributed systems",
        "title": "Bouncer API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.6"
    },
    "host": "localhost:5505",
    "basePath": "/",
    "paths": {
...
        "/watchdog/{name}": {
            "delete": {
                "description": "Remove a watchdog and clean up its resources",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Watchdog"
                ],
                "summary": "Delete a watchdog",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Watchdog name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Watchdog deleted successfully"
                    },
                    "404": {
                        "description": "Not Found - watchdog not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/watchdog/{name}/kick": {
            "get": {
                "description": "Reset the watchdog timer to prevent expiration",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Watchdog"
                ],
                "summary": "Reset watchdog timer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Watchdog name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 60000,
                        "description": "Time until expiration in milliseconds",
                        "name": "expires",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Watchdog timer reset successfully"
                    },
                    "400": {
                        "description": "Bad Request - invalid parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - watchdog not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/watchdog/{name}/wait": {
            "get": {
                "description": "Wait until the watchdog timer expires. Returns immediately if already expired.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Watchdog"
                ],
                "summary": "Wait for watchdog expiration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Watchdog name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": -1,
                        "description": "Maximum time to wait",
                        "name": "maxwait",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Watchdog expired or maxWait reached"
                    },
                    "400": {
                        "description": "Bad Request - invalid parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - watchdog not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "408": {
                        "description": "Request Timeout - maxWait exceeded",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
...
}